requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
pymongo>=4.6.0
python-dotenv>=1.0.0
requests>=2.31.0
numpy>=1.26.0
PyJWT>=2.7.0
cryptography>=41.0.0
